move parse_jsx_value to utils/react
